refactor:Update product update method for image change, Update delete method and check if the product exists

+++ src/main/java/com/example/myApp/controller/ProductController.java
@@ -10,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -61,15 +62,30 @@ public class ProductController {
                 .body(imageFile);
     }
 
-    @PutMapping("/products/update")
-    public void updateProduct(@RequestBody Product prod){
-        System.out.println(prod);
-        service.updateProduct(prod);
+
+    @PutMapping("/products/update/{id}")
+    public ResponseEntity<String> updateProduct(@PathVariable int id, @ModelAttribute Product prod,
+                              @RequestPart MultipartFile imageFile){
+
+        Product product = null;
+        try {
+            product = service.updateProduct(id, prod, imageFile);
+        } catch (IOException e) {
+           return new ResponseEntity<>("Failed to Update", HttpStatus.BAD_REQUEST);
+        }
+
+        if(product!=null) return new ResponseEntity<>("Updated", HttpStatus.OK);
+        else return new ResponseEntity<>("Failed to Update", HttpStatus.BAD_REQUEST);
     }
 
 
     @DeleteMapping("/products/{id}")
-    public void deleteProduct(@PathVariable int id){
-        service.deleteProduct(id);
+    public ResponseEntity<String> deleteProduct(@PathVariable int id){
+        Product prod = service.getProductById(id);
+        if(prod!=null) {
+            service.deleteProduct(id);
+            return new ResponseEntity<>("Deleted", HttpStatus.OK);
+        }
+        else return new ResponseEntity<>("Product Not Found", HttpStatus.NOT_FOUND);
     }
 }
